提交

+++ src/main/java/com/qiqilm/server/admin/im/ImApiImpl.java
@@ -188,7 +188,7 @@ public class ImApiImpl implements ImApi {
 	}
 
 	@Override
-	public boolean nospeakingT(String userId,int timeSec ) {
+	public boolean nospeakingT(String userId,Long timeSec ) {
 		final SetNoSpeaking kick = new SetNoSpeaking();
 		kick.setIdentifier( userId );
 		kick.setTimeSec(timeSec);
